Removing the PostService test class
